Moved AI code to model package
