+++ src/main/scala/org/esotericcode/reversi/gameengine/reversigameengine/CommandLineGame.scala
@@ -1,6 +1,7 @@
   package org.esotericcode.reversi.gameengine.reversigameengine
 
-  import org.esotericcode.reversi.gameengine.reversigameengine.model.{ImmutableReversiBoard, Node, ScoredNode}
+  import org.esotericcode.reversi.gameengine.reversigameengine.ai.{Node, ScoredNode}
+  import org.esotericcode.reversi.gameengine.reversigameengine.model.ImmutableReversiBoard
   import org.springframework.stereotype.Component
 
   import java.util
